+++ src/routes/index.svelte
@@ -1,51 +1,19 @@
 <script>
 	export const prerender = true;
-	// import { page } from '$app/stores';
-	// import { onMount } from 'svelte';
-	// let promise = undefined;
-	// let songs = Promise.resolve(undefined);
-	// const search = async (searchString) => {
-	// 	// var searchString = $page.params.searchString;
-	// 	const res = await fetch(`https://itunes.apple.com/earch?term=${searchString}&entity=song`, {
-	// 		// mode: 'no-cors' // 'cors' by default
-	// 	});
-	// 	var data = await data.json();
-	// 	console.log(data);
-	// 	if (res.ok) {
-	// 		return data;
-	// 	} else {
-	// 		throw new Error(data);
-	// 	}
-	// };
-	// function handleSubmit({ detail: { values, setSubmitting, resetForm } }) {
-	// 	setTimeout(() => {
-	// 		console.log(values);
-	// 		setSubmitting(false);
-	// 		resetForm({
-	// 			user: { email: 'test@user.com' }
-	// 		});
-	// 	}, 2000);
-	// }
+
 	import { goto } from '$app/navigation';
 	import SearchForm from '$lib/components/SearchForm.svelte';
-
-	// export const onSubmit = (e) => {
-	// 	const formData = new FormData(e.target);
-	// 	const searchString = formData.get('searchString');
-	// 	console.log(searchString);
-	// 	// goto(`${searchString}`);
-	// };
+	import { fade } from 'svelte/transition';
 
 	const clicked = () => {
-		console.log(`goto: ${searchString}`);
 		goto(`${searchString}`);
 	};
 
 	var searchString = '';
 </script>
 
-<div class="min-h-screen flex items-center justify-center px-16">
-	<div class="relative w-full max-w-lg">
+<div class="flex min-h-screen items-center justify-center px-16">
+	<div class="relative w-full max-w-xl" transition:fade={{ delay: 250, duration: 300 }}>
 		<SearchForm />
 
 		<!-- {#await songs then songs}