+++ CHANGELOG.md
@@ -1,2 +1,12 @@
+## [0.9.0] - 2022-03-21
+- ADD Font 
+- ADD Error page
+- CHG Design of result list and player
+- FIX Back button did not goto previous page
+
+
 ## [0.8.0] - 2022-03-21
 - Initial public release
+
+
+- Change log https://keepachangelog.com/en/1.0.0/